Respect user autostart settings.
Read `QUnit.config.autostart` after `test-helper` and friends have had a chance
to set the value.  Load the tests in either case, but only start them
automatically if `QUnit.config.autostart` has not been set to false.

This is particularly useful for users who want to debug tests in PhantomJS.