Correct "Doc test pane" to "Dock test pane"
This PR so far only changes the string displayed in the UI. The rest of the code still refers to this option as `doccontainer`. Should all these instances also be changed to `dockcontainer`?

As an aside, it seems to be called a container in one option ("Hide container") but as a test pane in another ("Doc test pane"). Should these be made consistent?