Fix issues with 1.2.0+ and ember-cli@2.2.0-beta.{1,2}.
ember-cli@2.2.0-beta.{1,2} had a bug that caused all legacy test files
to be appended in the wrong order. This meant when 1.2.0 was released of
ember-cli-qunit things completely broke (because QUnit was expected
before it was included in the file). The underlying issue was fixed upstream and released in
ember-cli@2.2.0-beta.3 (in
https://github.com/ember-cli/ember-cli/pull/5274).

This PR adds support for pulling qunit from `bower.json` on ember-cli's
< 2.2.0-beta.3 and uses NPM on ember-cli's >= 2.2.0-beta.3.