Prevent calling `QUnit.start` twice with jQuery 3.0.
Under jQuery 3.0 `$.ready` is always async (even if the document is already loaded). This is generally a great thing, however we were relying on the fact that it was sync (then adding our own async to ensure it was async).

This change prevents QUnit from auto-starting until our callback is completed.