Prevent clobbering custom `this.options.babel`.
Setting `this.options` inside of `included` to the options from
`lib/broccoli/ember-app` means that we will attempt to transpile
with the `options.babel` that the host app / addon is using.
Which, as of, 2.13+ that will be babel@6 and therefore incompatible
options with ember-cli-babel@5 (which this addon uses).

(cherry picked from commit 893d0ffb7ff460be8a275bec95e20c226417066b)