don't cover open Module drop-down
the `#ember-testing-container` has a z-index of 9999, which I'm guessing is random.  However, the Module dropdown in the Qunit header has a z-index of 99, so when passed tests are hidden and all tests pass, the `#ember-testing-container` covers the dropdown without this fix.  See https://monosnap.com/file/ouYQKdxc1mMzMpStK9SCp3iUnN6eLQ.  I changed `.full-screen` just for consistency, though it works properly either way.